+++ zotero-plugin/prefs.js
@@ -1,2 +1,3 @@
 pref("extensions.zpm.outputDir", "");
 pref("extensions.zpm.executablePath", "");
+pref("extensions.zpm.annotationLayout", "separate");